+++ test/src/derived_unit_tests/angle_tests.cc
@@ -1,0 +1,7 @@
+#include <catch.hpp>
+
+#include <SI/angle.h>
+
+/* This file does not yet contain any tests, as radiant and sterradiant are only
+ * typedefs without operators or conversions. However to ensure compileability
+ * this file includes the necessary headers*/